% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaangui-package.R
\docType{package}
\name{lavaangui-package}
\alias{lavaangui-package}
\title{lavaangui: Graphical User Interface with Integrated 'Diagrammer' for 'Lavaan'}
\description{
The lavaangui package provides a free, open-source graphical user interface
for the \pkg{lavaan} package. Models can be specified by drawing path diagrams
and fitted models can be visualized via interactive path diagrams.

The package offers two core functions:
\describe{
\item{\code{\link{lavaangui}}}{Launches a web application that allows users
to specify models by drawing path diagrams, fit them, and assess model fit.}
\item{\code{\link{plot_lavaan}}}{Creates interactive path diagrams from
fitted \pkg{lavaan} models. Nodes can be repositioned by dragging them
with the mouse.}
}

The web application can also be used without installing R at
\url{https://lavaangui.org/}.
}
\references{
Karch, J. D. (2025). lavaangui: A Web-Based Graphical Interface for
Specifying Lavaan Models by Drawing Path Diagrams. \emph{Structural Equation
Modeling: A Multidisciplinary Journal}, 1--12.
\doi{10.1080/10705511.2024.2420678}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lavaangui.org/}
  \item \url{https://github.com/karchjd/lavaangui}
  \item Report bugs at \url{https://github.com/karchjd/lavaangui/issues}
}

}
\author{
\strong{Maintainer}: Julian D. Karch \email{j.d.karch@fsw.leidenuniv.nl} (\href{https://orcid.org/0000-0002-1625-2822}{ORCID}) [copyright holder]

}
\keyword{internal}
