% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactive.R
\name{plot_interactive}
\alias{plot_interactive}
\title{Launch Interactive Plot for Lavaan Models}
\usage{
plot_interactive(fit, where = "gadget")
}
\arguments{
\item{fit}{The fit object that needs to be visualized interactively.}

\item{where}{A character string to specify where the interactive tool should be launched.
Takes one of two values: "gadget" or "browser". Default is "gadget".}
}
\description{
The \code{plot_interactive} function launches an interactive plotting tool
for visualizing lavaan models. The function can operate as either a
standalone gadget or within a browser, based on the 'where' argument.
}
\details{
This function leverages the \code{start_app} internal function to initialize the interactive
plot. It passes the \code{fit} object for visualization and sets the display location based on the
\code{where} parameter.
}
