% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_shiny.R
\name{lavaanGUI}
\alias{lavaanGUI}
\title{Shiny interface for lavaan}
\usage{
lavaanGUI(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Option will be passed on to \code{\link[shiny]{runApp}}}
}
\description{
This function calls a shiny interface for lavaan.
}

