% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_gui.R
\name{start_gui}
\alias{start_gui}
\title{Initialize and Run lavaan GUI Shiny Application}
\usage{
start_gui()
}
\description{
The \code{start_gui} function launches a Shiny application for lavaan GUI.
It sets up the UI from an HTML template and starts the application server.
}
\details{
This function performs the following tasks:
\enumerate{
\item Locates the "index.html" file within the "www" directory of the "lavaangui" package.
\item Adds a resource path for the "assets" directory.
\item Initializes and runs the Shiny application using the \code{lavaan_gui_server} server function.
}
}
