% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_gui.R
\name{start_gui}
\alias{start_gui}
\title{Start lavaangui Shiny Application}
\usage{
start_gui(fit = NULL)
}
\arguments{
\item{fit}{A lavaan model, as returned by the functions \code{lavaan}, \code{sem}, or \code{cfa} from the \code{lavaan} package.
If passed, lavaangui imports the model and data.}
}
\description{
The \code{start_gui} function launches the lavaangui Shiny application.
}
\details{
Currently, multiple-group models are not supported. However, you can create an
interactive plot of those models using \code{\link{plot_interactive}}
}
\examples{
\dontrun{
# Without importing lavaan model
start_gui()

# Importing a lavaan model
library(lavaan)
model <- ' 
  visual  =~ x1 + loadingx2*x2 + x3
  textual =~ x4 + x5 + x6
  speed   =~ x7 + x8 + x9
'
fit <- cfa(model, data = HolzingerSwineford1939)
start_gui(fit)
}
}
